pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "appointment-booking-system"
version = "0.1.0"
description = "A FastAPI-based appointment booking system with doctor and patient management"
requires-python = ">=3.10"
authors = [
    {name = "Anjali", email = "anjali@example.com"}
]
readme = "README.md"
license = {text = "MIT"}
keywords = ["fastapi", "appointment", "booking", "healthcare"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Healthcare Industry",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

dependencies = [
    "fastapi==0.125.0",
    "uvicorn==0.38.0",
    "pydantic==2.12.5",
    "pydantic-settings==2.12.0",
    "python-multipart==0.0.21",
    "sqlalchemy==2.0.45",
    "psycopg2-binary==2.9.11",
    "requests==2.32.5",
    "alembic==1.17.2",
    "python-dotenv==1.2.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]
test = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.24.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/appointment-booking-system"
Documentation = "https://github.com/yourusername/appointment-booking-system#readme"
Repository = "https://github.com/yourusername/appointment-booking-system.git"
Issues = "https://github.com/yourusername/appointment-booking-system/issues"

[tool.setuptools]
packages = ["app"]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312", "py313"]
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | migrations
)/
'''

[tool.isort]
profile = "black"
line_length = 100
skip_gitignore = true
known_first_party = ["app"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
markers = [
    "unit: unit tests",
    "integration: integration tests",
    "slow: slow tests",
]

